requirements.txt
langchain
langchain-google-genai
huggingface-hub
langchain_community
streamlit
python-dotenv